Player movement and speech bubbles
The player is split into two cause one needs to be the camera and the other needs to be able to move freely Would like to do it a different way but it works. The speech bubbles appear and disappear. I'm gonna make it easy to change what widget spawns by making the npc use a structure to get it's data or something.